add some global decorators
