+++ frontend/src/data-layers/regions/sidebar/RegionsSection.stories.tsx
@@ -1,19 +1,6 @@
 import { StoryObj, Meta } from '@storybook/react';
 
-import { atom, RecoilRoot } from 'recoil';
-import { useSyncRecoilState } from 'lib/recoil/sync-state';
-import { viewStateEffect } from 'state/view';
-import { useStateEffect } from 'lib/recoil/state-effects/use-state-effect';
-
 import { RegionsSection } from './RegionsSection';
-
-function recoilDecorator(Story) {
-  return (
-    <RecoilRoot>
-      <Story />
-    </RecoilRoot>
-  );
-}
 
 function fixedWidthDecorator(Story) {
   return (
@@ -23,32 +10,17 @@ function fixedWidthDecorator(Story) {
   );
 }
 
-const mockViewState = atom({
-  key: 'mockViewState',
-  default: 'exposure',
-});
-
-function SectionStyle({ children, view }) {
-  useSyncRecoilState(mockViewState, view);
-  useStateEffect(mockViewState, viewStateEffect);
-  return children;
-}
-
 const meta = {
   title: 'Sidebar/RegionsSection',
   component: RegionsSection,
-  decorators: [fixedWidthDecorator, recoilDecorator],
+  decorators: [fixedWidthDecorator],
 } as Meta;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {
-  decorators: [
-    (Story) => (
-      <SectionStyle view="exposure">
-        <Story />
-      </SectionStyle>
-    ),
-  ],
+  args: {
+    view: 'exposure',
+  },
 };
